remove the Tile.value field for perf gain
The value is only used for final display and erasing has_masks. It
can be recovered by using the try_order array and index of the Tile.
The reduction in space and removal of instructions needed to
maintain the field resulted in a performance gain:

An a/b test with the same random seed and the command `gen_ok 30000`
showed time elapsed dropping by 1/5 of the original value.

+++ cpp/src/solvent_cli/main.cpp
@@ -34,7 +34,7 @@ int main(const int argc, char const *const argv[]) {
 
 	user_order = (argc > 1) ? std::stoi(argv[1]) : SOLVENT_DEFAULT_ORDER;
 	if (argc > 2 && !std::string(argv[2]).empty()) {
-		srand_key = std::stoi(argv[2]);
+		srand_key = std::stoul(argv[2]);
 	} else {
 		srand_key = std::random_device()();
 	}
